Fixed 2D line name display in 3D scenes
Summary:
linename font should be the same as the one used for scene annotations.
It was set in setGeomID, but this is called before the display is added
to the scene. Now moved to the setGeometry function.

Perhaps we need 2 virtual fuctions in SurveyObject, initBeforeAdd and
initAfterAdd where these kinds of properties can be set.

OD-6276

Test Plan:
Added 2D line. Line is now clear and readable. It also responds to font and
color changes from the Scene properties window.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D17917