Fixed blank display of Steering data on plane displays
Summary:
Fixed incorrect return, when a new display element is added, it is
present in the uiVisPartServer instance but not yet in the uiODPlaneDataTreeItem
object.

[Jira: OD-6240]

Test Plan:
Added a new inline, selecting 'Steering data' at its first attribute.
Both components of the SteeringCube are now loaded automatically, as in v7