Improved output well selection
Summary

Added a well translator to support the reading and writing of HDF5 wells. That means that all the corresponding implementation is moved from odWellTranslator to hdfWellTranslator.
Cleanup of the WellDataIOProvider class: removed functions that do not need to be public, or implemented. WellDataIOProvider::canWrite is true by default (was not used, now it is used by WellDataIOProviderFactory::getWriteAccess)
Cleanup of WellTranslatorGroup and odWellTranslator
Added hdfWellTranslator
The WellTranslator implementations should take care of initializing the corresponding WellDataIOProvider implementations in their initClass static function

hdfWellTranslator: Always read support, write support if HDF5 is enabled (for wells)
odWellTranslator: Always read support, with support only if HDF5 is not enabled (for wells)
The default well translator is initialized as the one with both read and write support.

Moved the WellDataIOProvider implementations to the source file when odWellTranslator and hdfWellTranslator are implemented. They do not need to be placed anywhere else, or be made publicly accessible.
The 'default logs' were not correctly handled for the legacy wells (.defs file), for the rename, remove and copy well operations
The 'default logs' did not have an implementation for the HDF5 container, added it

HDF5:
Remove the macro to get the hdf5 file extension, there is already a static function for it. Added similar function for the icon identifier.
Many functions in the HDF5 well reader did not account for an IOObj object with no implementation (file missing). Based on the filename, this must be seen as a valid filepath to an HDF5 file.
Many public functions of the HDF5 well reader did not check if the reader (rdr_) was null, like in the case of a missing file, added the corresponding protections.
The renaming of an HDF5 well did not change the content of the HDF5 file: The root attribute 'Well name' needs to be updated. Not using a Well::Reader for that, directly using the (OD)HDF5 API for it.

I/O:
Bug fix in uiIOObjSel: The overwrite confirmation was most often only given once, and sometimes not at all. Fixed the bug, and added the saving of the user confirmation (per object).
uiIOObjTranslatorWriteOpts: deprecated two functions, replaced by standard usePar/fillPar, removed usage of the macro mDecluiIOObjTranslatorWriteOptsStdFns

Bug fix in uiIOObjSelWriteTranslator::selectedTranslator: if the GUI has only one translator enabled, and the default translator of the group is not 'userSelectable' (for read or write depending on the context), this wrong translator could still be returned. Now ensuring this does not happen. For instance a translator group with three translator templates, out of which two are not selectable (they do not support writing in a write context), the returned translator should be the only other translator left, not the first in the group.
uiIOObjSelWriteTranslator should be using fillObj/useObj when an IOObj object is available, and only fillPar/usePar in the absence of a contextual object.

Fixed similar missing overwrite prompt in the uiSeisWvltCreate class implementations, and uiWaveletExtraction

The copy well dialog did not allow overwriting existing wells (with a prompt), this is now supported.

Test Plan

Legacy and HDF5 wells can be displayed in the scene
Their data is available and correctly displayed in the well manager.

Well copier:
Legacy -> HDF5 (default mode, no option)
HDF5 -> HDF5 (default mode, no option)
Legacy -> Legacy (need to use the setting 'dTect.Use HDF5.Wells: No')
HDF5 -> Legacy (need to use the setting 'dTect.Use HDF5.Wells: No')