Update of IOStream::implReadOnly [devel]
Summary:
For the translator-based IOObj implementations, IOObj::isUserSelectable(forread=false)
is a synonym of seeing the IOObj as read-only: The translator specifies
that writing to this object is either not supported, or should not be done.

Test plan:
In od7.0, hdf5 wells are now seen as read-only, without having to read
or change the read-only flag of the file implementation.
Direct links to SEG-Y, ..., still behave as before.