Fixed bugs in uiioobjsel
The overwrite confirmation was most often only given once, and sometimes not at all. Fixed the bug.

Bug fix in uiIOObjSelWriteTranslator::selectedTranslator: if the GUI has only one translator enabled, and the default translator of the group is not 'userSelectable' (for read or write depending on the context), this wrong translator could still be returned. Now ensuring this does not happen. For instance a translator group with three translator templates, out of which two are not selectable (they do not support writing in a write context), the returned translator should be the only other translator left, not the first in the group.
uiIOObjSelWriteTranslator should be using fillObj/useObj when an IOObj object is available, and only fill(Par)/use(Par) in the absence of a contextual object.