Added HDF5 well reader
Summary:
Added a well translator to support the reading of HDF5 wells. That means that all the corresponding implementation is moved from odWellTranslator to hdfWellTranslator.
Added hdfWellTranslator
The WellTranslator implementations should take care of initializing the corresponding WellDataIOProvider implementations in their initClass static function

odWellTranslator: Read/write, default (no change)
hdfWellTranslator: Read-only support, providing that the HDF5 plugin is available.
The default well translator is initialized as the one with both read and write support, thus the legacy format in this branch.

Moved the WellDataIOProvider implementations to the source file when odWellTranslator and hdfWellTranslator are implemented. They do not need to be placed anywhere else, or be made publicly accessible.
The 'default logs' were not correctly handled for the legacy wells (.defs file), for the rename, remove and copy well operations

All hdf5-based implementations are placed in the source files, to allow any future modification is required.

Test Plan:
Legacy and HDF5 wells can be displayed in the scene
Their data is available and correctly displayed in the well manager.

Well copier:
Legacy -> Legacy
HDF5 -> Legacy

Reviewers: nanne, arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D17936