Fixed issues with the WindowsTerminal prompt
Summary:
When launching the windows terminal application without using the custom python activator, thus using the standard conda activation scripts, the OpendTect environment was not forwarded to the windows terminal. As a result, this terminal did not have the pythonpath environment variable set, and the path to the python interprer was also missing.
Added the '--inheritEnvironment' command line flag to fix it.
Also adjusted the default arguments used when starting a terminal prompt from the python access class, to annonate it using the virtual environmnent name.
Updated the C++ syntax to create the temporary script, to avoid a strange "PIDRET=" in the console output when the process ID of the background application cannot be retrieved, as it is the case when launching the WindowsTerminal application.

Test Plan:
With and without custom python activator application:
Started the (legacy) command prompt, powershell and WindowsTerminal applications from the toolbar button, tested that
in all cases the python interpreter could be started, and has access to the python modules odpy and dgbpy.
Also checked that the virtual environment name was used in the annotations (not available for powershell).
Checked that the bokeh server was correctly started in the background,
all of that without spurious print in the standard output/error.

Reviewers: nanne

Reviewed By: nanne

Subscribers: wayne

Differential Revision: https://phabricator.dgbes.com/D17958