Fixed conversion wchar_t to BufferString conversion [devel]
Summary:
BufferString::BufferString, BufferString::set and BufferString::add would
result in a "Yes" output string when using a wchar_t input argument,
because if would use the boolean value of the template to run the implementation.
To avoid it, defining additional functions (this is an exception), only available on Windows,
as the wchar_t type is not defined on other platforms. This is in line with what Qt is doing for
its QChar class.

Test Plan: Expanded the test program. Run successfully on Windows.

Reviewers: khushnood

Reviewed By: khushnood

Subscribers: nanne, raman

Differential Revision: https://phabricator.dgbes.com/D17967