Fixed potential crash in Wavelet::getIOObj
The crash occurs often, but not always, in test_synthseis, on Windows.
Could not reproduce the crash after removing the copy operation of the IOObjContent object.
When crashing, the content copy constructor argument, which is a static singleton, is not valid (corrupted), although the destructor of the WaveletTranslatorGroup has not been hit (not even once deleted, as expected).
The function is called many times in a multithreaded execution, which suggests
that there is still a potential thread safety issue in the IOObjContext class,
or in the IOPar class, which the class really occurs (because of its hashmap).