Corrected usage of add_custom_command in ODBreakPadUtils.cmake
Summary:
Policy CMP0175 warns that the 'DEPENDS' keyword should not be used when
calling the cmake function add_custom_command with the 'TARGET' keyword.
This policy was introduced in cmake 3.31.
The 'DEPENDS' is not required anyway, as its value is the same as the
value provided to the 'TARGET' keyword.
Hence the corresponding line can just be removed, for the same result.

Test plan:
Re-built the OpendTect targets, symbols remain created.