Added a wrapper to the CallBacker class to enable Lambda expressions
Summary:
It is not possible to directly use Lambda expressions using the CallBack, CallBacker and Notifier classes,
as they are too strongly typed.
The LambdaCallBacker derived class enables it. The class is required to manage the function pointer.
It allows retrieving the corresponding CallBack object, attaching and detaching to Notifier objects,
manually (NotifierAccess::notify and NotifierAccess::remove), or automatically (LambdaCallBacker::attachCB
and LambdaCallBacker::detachCB).

See example in the test program

Test Plan: Made a specific function in test_callback.

Reviewers: wayne

Reviewed By: wayne

Subscribers: raman, nanne

Differential Revision: https://phabricator.dgbes.com/D17976