Small changes in the initialization of uiODMain
Summary

Removed the construction of an ApplicationData object to support the creation of dialogs
prior to the execution of the main event loop. This should not be required.
Made the splash screen a local object (no need to store it in a PtrMan). Added explicit call
to its finish function, as per the Qt documentation: The splash screen will be removed on the
first call of the show() function for the given argument, the od_main main window.

Test Plan

Splash screen is displayed correctly, in the free and Pro version of the software
(was already good for the free software).