Added copy constructors based on OD::String references [devel]
Summary:
In some cases, calling a function that requires a 'const char*' parameter, by providing as argument
a static function to a potentially empty static string buffer, simply does not work on
the Windows platform, because the static function is not entirely executed while
evaluating the argument of the function.
The only possibility to get this safe is to stop returning a 'const char*' buffer for these types
of functions. One could use a StringView return type, but this would still create on-the-fly a
StringView object, which can be skipped by returning the OD::String const reference static string.
Added few constructors and set functions that can directly pass the reference to the string,
hence the direct return type of these functions, to avoid having the pollute to code of the callee with a '.buf()'
function call.

Test Plan:
See the test program function testStaticString() for a typical use case.
It is not fully clear why this simple test program works on windows too, whereas similar structures in plugins sometime fail.

Reviewers: khushnood

Reviewed By: khushnood

Subscribers: nanne, raman

Differential Revision: https://phabricator.dgbes.com/D17979