Fixed initialization of od_main with Splash screen [devel]
Summary:
Removed the construction of an ApplicationData object to support the creation of dialogs
prior to the execution of the main event loop. This should not be required.
Made the splash screen a local object (no need to store it in a PtrMan). Added explicit call
to its finish function, as per the Qt documentation: The splash screen will be removed on the
first call of the show() function for the given argument, the od_main main window.

Reworked entirely the flow to make it Qt compliant: There should be a single call to QApplication::exec.

As a result, one must implement and chain into a sequence several lambdas:
1- The first retrieves the product selection window, and 'shows' it if there is one to show.
2- The second waits until the product selection window is closed, and triggers the startup sequence
3- The third, the startup sequence, sets the splash screen and 'shows' the uiODMain main window.

Both the dialog and the main window must remain valid outside of the scope of the ODMain extern function.
PtrMan<T>& do not work when capturing the references in the lambdas, in a threaded context,
hence the use of std::unique_ptr and std::make_shared containers, which have all the required support for it.

The proposed flow is utterly complex, but required since Qt does not natively support running consectively
two QMainWindow, and interleave a QSplashScreen in between. The flow would become much simplier if the first
window would be a real QDialog, not a uiDialog inheriting from QMainWindow. In that case this non-modal
QDialog could be executed entirely in the scope of the product selection function, and would not need to be
exported.

Test Plan:
Tested on Windows and Linux: OpendTect Free, OpendTect Pro with and without selection window.
In all cases the splash screen is now correctly displayed after the product selection window (and its deletion
if applicable), and before the uiODMain main window is displayed.
OpendTect starts correctly, ends with the correct exit code, no memory error reported (Windows was very picky on that).