Regional marker manager can now be run from Advance Well Manager
Summary:
Regional markers are now saved as separate stratigraphy file in wells.regm in models folder of a survey. Thus they are unique to a survey. These are then read and stored in Strat::RegMarkerSet class as Strat::RegMarker objects. Such object is the same as Strat::Level object except the id assigned to it now starts at 10000 and that they can be accessed as reference to stored object and not as a copy.

Strat::LevelSet class also stores these objects as a copy, but their id cannot be modified in this class. Neither can one be added to Strat::LevelSet. If you need to add a RegMarker, it must be added using Strat::RegMarkerSet class, which then updates LevelSet class too. The same regional markers are also made available to all the levelsets created in a session and are thus independent of Stratigraphy model chosen.

On the UI side, a new dialog has been created. It can be launched from Advance Well Manager where markers can be assigned as regional markers, or grouped together under a regional marker.

Test Plan: tested in Advance Well Manager in multiple surveys with no stratigraphy levels, one stratigrphy level and multiple stratigraphy level.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D17993