Now range volume subsel fields snap to valid line number for hsamp.step_ > 1
Summary:
If the horizontal sampling step is greater than one, then uiSpinBoxs in volume subselection dialog can only be toggled in the multiple of that step for inlines and crosslines. This way the user cannot set an inline or crossline that is not start + x*step_.

for ex. inline.start_ = 100, inline.step_ = 2, then only 100, 102, 104  and so on will be accepted in the spinbox.

Additionally, if step field has been chosen, then step_ value can only be incremeneted and not decreased.

Test Plan: tested in various seis manager dialogs.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, nanne

Differential Revision: https://phabricator.dgbes.com/D17999