Fix for Depth to time conversion using a velocity model
Summary: In a depth survey, conversion of seismic data to time did not work. The VolumeOfInterest is in time, whereas the velocity data is in depth. The reader (and the translator) gets a SelData with z range in time and therefore fails to read anything. Now setting the SelData z range to that of the velocity volume (depth) fixes the problem. It already works in v2025 because the SeisTrcTranslator is handling the ZDomain differences.

Test Plan: Converted a depth volume to time on-the-fly (in a time scene) and in batch. Results are as expected.

Reviewers: nanne, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D18008