Added compatibility to MS-DOS zip archives
Summary:
The ZipHandler was not capable of unpacking zip archives created by the
native windows compressor (from the contextual menu in the file explorer).
This is caused by a lack of support for the data descriptor, which is an optional
structure that is written after the compressed file data, then before the local file
header of the next file.
Also the files with such a data descriptor have a zero size compressed and uncompressed data
in the local directory, thus one needs to take care of not overwriting the values read from the
central directory.
The ZipFileInfo will not keep track of these attributes, but they will only be set during
unpacking operations, for instance the ZipArchiveInfo won't be able to access this information,
as it is only reading the information from the central directory.

Since one cannot add class members to the ZipFileInfo class, and to avoid introducing a HiddenParam,
simply setting the read position at the start of each ZipHandler::readLocalFileHeader, if needed.

Note that this is not a regression as compared to v6/v7, this has never been implemented, and was probably never seen because neither the Unix zip tool nor 7-Zip use that data descriptor when producing zip archives.
For the record, these native windows archives do store correctly the timestamps and time zone information, thus these are correctly set when unpacking such archives. But they do not store nicely the windows file attributes, thus a local cycle compression/uncompression on the same window machine may induce loss of information (as the files are marked as being produced on Unix, with Unix-like file permissions).

https://en.wikipedia.org/wiki/ZIP_(file_format)#Data_descriptor

Test Plan:
The test program test_zip still is OK.
Was able to uncompress a zip file created on Windows, that contains 6 files.

Reviewers: nanne, arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D18023