Added visSurvey::WellDisplay::trackDataRequest_ [devel]
Summary:
The well display positions used in the 3D vizualisation are coming from visSurvey::WellDisplay::getTrackPos,
which fills a TypeSet<Coord3> object. This is not only done once at the construction of the object, but on
each fullRedraw.
As a result, to be able to use a different set of values, one needs to insert and use a signal that will fill
this set of positions. If filled externally, then the implementation within visSurvey::WellDisplay::getTrackPos
should not be used.

Use case:
Attach a notifier to uiVisPartServer::objectAdded.
In the implementation of the corresponding callback function:
1- Dynamic cast the added object to a visSurvey::WellDisplay. Return if it is a different type.
2- Implement your callback function: Fill the TypeSet<Coord3> from the capsule.
3- Attach your callback function to the visSurvey::WellDisplay::trackDataRequest_ notifier of this object.
4- Only once, right after having attached your callback function to the visSurvey::WellDisplay::trackDataRequest_
notifier, call the public function visSurvey::WellDisplay::redraw(). This is because the well display has already been
created (but not yet displayed on screen) when uiVisPartServer::objectAdded is called.

Test Plan: compiles and links, to be used from plugins.

Reviewers: nanne, raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com/D18024