Restored timing of ApplicationData::applicationToBeStarted() signal
Summary
The purpose of this signal is to inform modules and plugins that that top level widget has been constructed, and that it is displayed for the first time.
It should thus not be triggered before all plugins have been loaded, before the implementation of the top level widget (uiMainWin) is constructed or finalized.

For this reason, the trigger cannot stay in uiMain::exec, this is too soon, it needs to be moved to uiMainWin::show.

Test Plan
the trigger is now correctly caught by all plugins that need it. All menu items from plugins are back.