Added Task::progressFactor virtual function [devel]
Summary:
For some tasks, the nrDone/totalNr to be displayed needs to be displayed
in a certain magnitude. For instance if the progress needs to be
reported in kilobytes or megabytes, but the work unit is in bytes.
Using a separate scale factor allows removing the scale factor from the
nrDone/totalNr implementations.
The main advantage is to avoid truncations to nrDone/totalNr because of
the scale factor.

Test Plan:
Coverered by the following test programs:
test_file
test_zip
test_networkaccess

No visible change in the progress report seen in the log file or
uiTaskRunner.