Sequential task doesn't call nextStep if totalNr is empty
Summary:
If number of items to be worked on is empty, then nextStep should not be called, to be inline
with the behaviour of the ParallelTask class: nrDone and totalNr must be set when leaving the constructor,
to the final value if possible, to a guessed value otherwise, more than zero for totalNr().

To guarantee that nextStep will be called, once can thus for instance set totalNr() to 1 in the constructor,
and revise the value in the doPrepare function.

Test Plan:
Several tasks have already been converted. Expect to find more,
but for now all test programs run successfully.

Behaviour in release mode does not change. Can be disable by setting the
environment variable 'DTECT_TRACK_SEQUENTIALTASKS' to 'No'
Should NOT be ported in a release branch.