Rework of GenericTransformND task [devel]
Summary:
Updated usage of SequentialTask to the current standards.

GenericTransformND should be protected, this is a base class.
Implemented: nrDone, totalNr, doPrepare, doFinish. curdim_, the implementation of nrDone, becomes private, it does not need to be set to -1 each time an input or parameter changes.

Test Plan: Tested from test_dip_steering

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com/D18032