Improvements in Table::Converter task [devel]
Summary:
  - The task should be a SequentialTask, it does not use the Executor specific API
    a quick implementation for it (should be made parallel eventually)
  - In some cases, the maximum number of lines to be read is known, this
    information can now be provided to the constructor, and will make
    the task run faster when reading few lines
  - When used from Table::AscIO, the Table::Converter should know how
    many (header) lines have already been read when executing Table::AscIO::getNextBodyVals,
    thus added a public function for it.
  - The converter must know the total number of lines in the file, made
  - The execution functions shoud be private. Added a specific access to AscIO

Test plan:
Used import of well locations GUI for testing, and the examine button
of the file input selection (which uses od_FileBrowser).
This shows an important flaw in the examine window: It does not know if
it is dealing with a blank separated file or CSV file. To be improved in
a different revision.