Small update in WSImportHandler [devel]
Summary:
The blank separated ASCII files, using the same logic as for CSV files:
If a line separator is found within a field (quoted text), ignoring this
line separator, and replacing it with a space.

Test plan:
Tested from well locations import. The real number of header lines, 4,
is reduced to 2 after parsing, as it was in the original excel sheet.
The examine window shows a 2 lines header in both cases.