Added od_istream::getAsciiNrLines [devel]
Summary:
There was no tool that could effeciently tell us how many lines are present in a text file.
Made a very optimized parallel task for this purpose, used in the constructor of the
Table::Converter class.
Can be used as a template for pre-scanning text files for other purposes, for instance we
really miss such a task that automatically determines if a file is comma separated or
space/tab separated.

Test Plan:
Read a text file with almost 15 million lines, 650MB large.
The AWK processor needs about 1.5 sec to return the number of lines.
The previous simple sequential implementation runs in 4700ms.
The new parallel task implementation runs in about 450ms (machine with 20 CPUs).
It runs in 4ms when limiting the scope to the first 10000 lines. This allows the TableConverter
class to process faster when its only purpose is to parse the file header, or when it is
being used by the ascii file examine window.

Reviewers: nanne

Reviewed By: nanne

Subscribers: alexander, raman

Differential Revision: https://phabricator.dgbes.com/D18041