Switching to xcb_egl for some of the OpendTect applications [devel]
Summary:
Trying to fix the issue where the OpendTect installer does not start on Ubuntu 24.04, when
not using an Nvidia GPU.
This fix should improve the compatibility on Linux/X11 exclusively: All Qt apps use the xcb
platform by default, and this is the only platform that would work on Linux/X11 anyway.
For this platform, switching the rendering to 'xcb_egl' for all GUI applications that do not
use OpenGL (thus 3D) rendering, instead of the default 'xcb_glx', only if the OpenGL renderer
is not NVidia. Thus this applies to all GUI apps, except od_main.

Test Plan:
Tested as much as possible from Alma Linux 9.5. Awaiting more tests