Updated deployment of the Qt plugins [devel]
Summary:
The cmake variable QT_REQ_PLUGINS is no longer a list of plugin (folder)
names, but a list of files within these folders. This is the only way to
safely include only some of the files in the plugins folders, especially
on Windows, where the plugins have debug and release runtime libraries.
The list is made out of the libraries that apply to all build
configurations (Unix), or the release configurations (Windows).

Main differences in the deployed files:
- Not all platforms are installed/packages, only a subset
- The PDB files of the Qt plugins in release mode are only installed in RelWithDebInfo configuration
- The PDB files of the Qt plugins in debug mode are only installed in Debug configuration
- The base package only contains release runtime libraries of the Qt plugins
- The devel package on Windows now contains the debug runtime libraries of the Qt plugins
Not changed: the base and devel packages do not contain the PDB files of the Qt plugins

Test plan:
Built and installed the project on Windows and Linux, created the
packages, checked their content.