Terminating stream if a process is terminated in od_progressviewer
Summary:
The logfile would be locked as the stream was still open after terminating the process. This meant that the process couldn't be restarted with same logfile without closing progress viewer window.

This revision closes the stream if process is terminated and releases the logfile.

Test Plan: tested using a batch program by terminating it before finishing, then relaunching the program without closing the first progress viewer.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D18063