Moved most of the processing for wavelet extrction to static functions
Summary: Created new static functions which are container independent for trace input. These functions take the relevant sections of a trace as an Array1D object and this can be called for SeisTrc trace or a DataPack or some other object as long as the trc data can be accessed as Array1D object.

Test Plan: Extracted Wavelet and compared to an old verison for the same parameters. Wavelets are same.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D18079