Fix for Bayes Classification Output Dialog
Summary:
If the user pressed abort or attempted or attempted go back and forth in
the dialog, there would be a new window that opened up, and if the user
attempted to exit, would crash OpendTect.

Test Plan:
Compiled and ran OpendTect. Attempted to run the Bayes dialog
with different options, aborting the process, and going back and forth
in the dialog wizard.

Reviewers: nanne, raman

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D18094