Now writing uwi in .omf for wells
Summary:
Since all the well writers are derived from WriteAccess, the WriteAccess when created now checks whether the well exists or not. If it doesn't, and has a valid UWI, will write this in the .omf file.

The writing is done using a static function in Well namespace. This function should be used to update the "omf" entries when UWI's are edited in opendtect.

Test Plan: tested by importing wells from owx and Petrel.

Reviewers: raman, arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D18104