Enabling and disabling position slide for RandomLines after manipulation
Summary:
The sliders for RandomLines are available only if it has two nodes and is along an inline or crossline. If the random line is manipulated and the aforementioned criteria is not met, position sliders should be hidden, and vice versa.

This revision takes care of that.

Test Plan: tested in 3D scene.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D18125