Fix RGB<->HSV color conversion error
Summary:
Found RGB->HSV color conversion included a truncation error for the hue component when
it was converted from int to unsigned char. Hue xomponent now interpreted as a value
in the range 0-255 corresponding to hue angles of 0-360 degrees.
Added distinctColors function to get array of colors with same saturation and brightness
but different hues.
Added test program for RGB<->HSV conversion and distinctColors and complimentaryColors
methods.

[Jira: OD-6177]

Test Plan:
Displayed wells in the 3D without stored display properties. Color is
the same as in the main branch