Fixed empty well track display in the 3D scene
Summary:
The function Well::DisplayProperties::ensureColorContrastWith was no longer hit in
uiODWellParentTreeItem::handleSubMenu, as the color comparison was not correct: The color of
Well::DisplayProperties::track_ is set with the default color in the
constructor of Well::DisplayProperties.

[Jira: OD-6527]

Test plan:
Loaded a well which has an empty display properties file in the scene.
Default track display color is red, observed a dark blue color used when
displaying this well track, instead of black.