Position dlg for volume items shouldn't have scroll and apply
Summary:
For Volume item, Scroll and Apply Button were visible in the dialog. These buttons did nothing. Most importatnly, the buttons were in the wrong places, and totally messed up the layout.

This revsion disables the display of those button in this specific case.

Test Plan: tested in Position Dialog in Z-transformed scene

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D18128