Enabled file operations on long windows file paths
Summary:
The following operations did not work on files/folders with a long path (>MAX_PATH=260):
Creating an input/output stream
Getting/setting file permissions
Getting/setting time stamps
Creating/remove files and directories.

The solution is to add a prefix to the wide string send to the WIN32 API functions,
which is captured in the implementation of FilePath::getWLongPath.
This should fix at least the unzipping errors we get when unpacking the Python environments on Windows, within certain directories.
Note that it is not required to edit the registry for this support to be added.

Also fixed some minor logic errors in the ZipHandler.

Reference: https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation?tabs=registry

Test Plan:
Expanded test program test_zip, which is a good test for file-based operations,
as it creates and removes files and directories, gets and sets permissions and timestamps.

Also zipped the test folder using both 7-zip and the native Windows zip context menu.
Ensured both could be unpacked without error.