Bugfix for Extra Point Added When Moving Left Extreme Anchor in the CT
Summary:
Changed a condition in the function display that would add a point
to the color table when moving the left extreme anchor. This
behavior does not occur with the right extreme anchor.

Test Plan:
Compiled and ran ODT. Moved the left extreme anchor to see
if a point would be added. Nothing was added. Right extreme anchor did
not have that problem, but tested it anyway, no problems found.

Reviewers: nanne, raman

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D18144