Fixed data being displayed on sections outside the range
Summary: When a section with displayed texture was moved outside the range of data availability, the section moved but still showed the data (texture) from the previous section. Setting a null datapack does clear the texture and fixes this issue.

Test Plan: Tested by moving Inl/Crl/Zslices outside the range of a small subcube. It shows an emty display with the warning sign in the tree. Moving the section back into the range displays the data correctly again.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D18149