Update section displays when pos is changed without using the slider
Summary: When the slider is not active, position can be changed either using the arrow keys or by simply entering a position by hand. Then the display must update.

Test Plan: Tested moving the slice using the slider, arrow keys, dragger and the position field for both preloaded and non-preloaded data. The behaviour is as expected.

Reviewers: nanne, khushnood

Reviewed By: khushnood

Differential Revision: https://phabricator.dgbes.com/D18150