Fixed a bug which would crash ODT when deleting a point on the ColorTable Transparency Canvas, implemented some  Anchor Changes
Summary:
Fixed a bug that would crash ODT when attempting to delete a point on
the transparency canvas

Added a few changes that did not make it to the 8.0 branch for the
anchors

Test Plan:
Compiled and ran ODT. Successfully deleted and added points
on the ColorTable Transparency Canvas.

Deleted and added anchors successfully.

Reviewers: nanne, raman

Reviewed By: nanne

Subscribers: arnaud, hadyan

Differential Revision: https://phabricator.dgbes.com/D18127