Fixed slowness in GenericTransformND
Summary:
Test setup() function does not need to be run on each execution,
it is costly and only needs to be run when the key parameters have been
changed.
For that reason, transforms_, transforms1dstarts_ and nr1dtransforms_
should not be deleted in doFinish, such that they are already allocated
for the next execution.

Test Plan:
test_dip_steering_vb is as fast as in od8.0 now on Windows too.