Reworked SimpleTextStreamProgressMeter output [devel]
Summary:
Some tasks have on purpose no name and/or no message, which should
be used to limit the verbosity of any ProgressMeter.
Also, the SimpleTextStreamProgressMeter was not getting the uiMessage()
of the task to display a contextual message.

Test Plan:
Checked log output from a Volume Processing. Looking good.